gcc 9 fixes
UPSTREAM: Kbuild: suppress packed-not-aligned warning for default setting only

gcc-8 reports many -Wpacked-not-aligned warnings. The below are some
examples.

./include/linux/ceph/msgr.h:67:1: warning: alignment 1 of 'struct
ceph_entity_addr' is less than 8 [-Wpacked-not-aligned]
 } __attribute__ ((packed));

./include/linux/ceph/msgr.h:67:1: warning: alignment 1 of 'struct
ceph_entity_addr' is less than 8 [-Wpacked-not-aligned]
 } __attribute__ ((packed));

./include/linux/ceph/msgr.h:67:1: warning: alignment 1 of 'struct
ceph_entity_addr' is less than 8 [-Wpacked-not-aligned]
 } __attribute__ ((packed));

This patch suppresses this kind of warnings for default setting.

Signed-off-by: Xiongfeng Wang <xiongfeng.wang@linaro.org>
Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
(cherry picked from commit 321cb0308a9e76841394b4bbab6a1107cfedbae0)
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: Akhil Narang <akhilnarang.1999@gmail.com>
Kbuild: Adjust disabled warnings for GCC 8.1.1

-Wattribute-alias spams due to the syscall macros. There was a patch[1]
to disable this warning just for the syscall macros but there were not
many comments and Arnd mentions later[2] that that version needs to be
updated.

-Wstringop-truncation should be hidden behind W=1 per Arnd[2].

-Wformat-overflow was removed in f86ee348ef8d ("wahoo: Fix all
outstanding GCC warnings") but it shouldn't have been because of
83863589806d ("disable new gcc-7.1.1 warnings for now").

[1]: https://lkml.org/lkml/2017/12/5/581
[2]: https://www.spinics.net/lists/stable/msg242639.html

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: Akhil Narang <akhilnarang.1999@gmail.com>
Makefile: Don't disable -Wmaybe-uninitialized globally

This is a very useful warning that frequently spots out blatantly incorrect
code. Let's not disable it.

Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: Akhil Narang <akhilnarang.1999@gmail.com>
Makefile: Disable '-Wmaybe-uninitialized' for GCC <= 5.0

There are a few warnings that appear to be false positives as they are
not present with both GCC 7.3.x and 8.2.x.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: Akhil Narang <akhilnarang.1999@gmail.com>
BACKPORT: exec: avoid gcc-8 warning for get_task_comm

gcc-8 warns about using strncpy() with the source size as the limit:

  fs/exec.c:1223:32: error: argument to 'sizeof' in 'strncpy' call is the same expression as the source; did you mean to use the size of the destination? [-Werror=sizeof-pointer-memaccess]

This is indeed slightly suspicious, as it protects us from source
arguments without NUL-termination, but does not guarantee that the
destination is terminated.

This keeps the strncpy() to ensure we have properly padded target
buffer, but ensures that we use the correct length, by passing the
actual length of the destination buffer as well as adding a build-time
check to ensure it is exactly TASK_COMM_LEN.

There are only 23 callsites which I all reviewed to ensure this is
currently the case.  We could get away with doing only the check or
passing the right length, but it doesn't hurt to do both.

Link: http://lkml.kernel.org/r/20171205151724.1764896-1-arnd@arndb.de
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Suggested-by: Kees Cook <keescook@chromium.org>
Acked-by: Kees Cook <keescook@chromium.org>
Acked-by: Ingo Molnar <mingo@kernel.org>
Cc: Alexander Viro <viro@zeniv.linux.org.uk>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Serge Hallyn <serge@hallyn.com>
Cc: James Morris <james.l.morris@oracle.com>
Cc: Aleksa Sarai <asarai@suse.de>
Cc: "Eric W. Biederman" <ebiederm@xmission.com>
Cc: Frederic Weisbecker <frederic@kernel.org>
Cc: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
(cherry picked from commit 3756f6401c302617c5e091081ca4d26ab604bec5)
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: Akhil Narang <akhilnarang.1999@gmail.com>
whyred: misc: fix mismerge of dd481e59528e54c6103f6ed82961a057043ca283

Signed-off-by: Akhil Narang <akhilnarang.1999@gmail.com>
qcacld-3.0: fix GCC 9.0 warnings

Signed-off-by: Yaroslav Furman <yaro330@gmail.com>
Signed-off-by: Akhil Narang <akhilnarang.1999@gmail.com>
codecs: disable -Wshift-overflow warning

Change-Id: Ia620daf7908ab6aff6de21c1badafd7313128920
Signed-off-by: Akhil Narang <akhilnarang.1999@gmail.com>

# Conflicts:
#	Makefile