revert: "msm: net: API to send list of SKB's to network driver
Existing network stack API's which are used to transmit
data to the network drivers send one SKB at a time. There
can be usecases where network driver expects a list of skb's
from the network stack. For example it will be needed in cases
where aggregation is needed before giving packets to HW. Create
a new API to send list of skb's from the network stack to
network driver. Re-use the existing functionality of
__dev_queue_xmit().

Change-Id: I67bd081c92eda63d3712d29f3b5bee21cde798ab
Acked-by: Chaitanya Pratapa <cpratapa@qti.qualcomm.com>
Signed-off-by: Mohammed Javid <mjavid@codeaurora.org>
 951d1abd228f1c2e311c017f54aa2f7fe6a038de"